+++ src/app/app-routing.module.ts
@@ -1,3 +1,4 @@
+import { NgModule } from '@angular/core';
 import { SignUpComponent } from './component/user/auth/sign-up/sign-up.component';
 import { AuthComponent } from './component/user/auth/auth.component';
 import { SignInComponent } from './component/user/auth/sign-in/sign-in.component';
@@ -10,14 +11,14 @@ import { FilterComponent } from './component/filter/filter.component';
 import { RestoreFormComponent } from './component/user/restore-form/restore-form.component';
 import { FeedbacksComponent } from './component/admin/feedbacks/feedbacks.component';
 import { UserHabitPageComponent } from './component/user/habit/user-habit-page/user-habit-page.component';
-import { Routes } from '@angular/router';
+import { Routes, RouterModule } from '@angular/router';
 import { HomepageComponent } from './component/general/homepage/homepage/homepage.component';
 import { AuthPageGuardService } from './service/route-guards/auth-page-guard.service';
 import { HomePageGuardService } from './service/route-guards/home-page-guard.service';
 import { AppComponent } from './app.component';
 import { EcoNewsComponent } from './component/general/eco-news/eco-news.component'
 
-export const router: Routes = [
+export const routes: Routes = [
   {
     path: '',
     component: AppComponent,
@@ -65,3 +66,14 @@ export const router: Routes = [
     component: EcoNewsComponent
   }
 ];
+
+@NgModule({
+    imports: [
+        RouterModule.forRoot(routes)
+    ],
+    exports: [
+        RouterModule
+    ]
+})
+
+export class AppRoutingModule {}
