+++ src/app/component/admin/feedbacks/dialog-photo/dialog-photo.component.ts
@@ -1,6 +1,6 @@
 import {Component, Inject, OnInit} from '@angular/core';
 import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material';
-import {ModalService} from '../../../user/_modal/modal.service';
+import {ModalService} from '../../../core/_modal/modal.service';
 
 @Component({
   selector: 'app-dialog-photo',
