+++ src/app/component/core/_modal/modal.component.ts
