+++ src/app/component/core/footer/footer.component.css
