+++ src/app/component/core/footer/footer.component.spec.ts
