+++ src/app/component/core/footer/footer.component.ts
