+++ src/app/component/core/photo-upload/photo-upload.component.scss
